WIP: Break up: Separate TOML collation and filesystem functions
