+++ topiary-cli/src/configuration/fs.rs
@@ -1,0 +1,19 @@
+use directories::ProjectDirs;
+
+/// Find the OS-specific configuration directory
+pub fn find_os_configuration_dir() -> PathBuf {
+    ProjectDirs::from("", "", "topiary")
+        .expect("Could not access the OS's Home directory")
+        .config_dir()
+        .to_path_buf()
+}
+
+/// Ascend the directory hierarchy, starting from the current working directory, in search of the
+/// nearest `.topiary` configuration directory
+pub fn find_workspace_configuration_dir() -> Option<PathBuf> {
+    current_dir()
+        .expect("Could not get current working directory")
+        .ancestors()
+        .map(|path| path.join(".topiary"))
+        .find(|path| path.exists())
+}
